app/requirements.txt
pillow==9.0.1
tflite-runtime